+++ src/quilt4/namespace.rs
@@ -4,7 +4,6 @@
 //! 
 
 use super::{
-    client::Client,
     domain::Domain,
     upath::UPath,
     manifest::Manifest4,
@@ -22,11 +21,6 @@ impl Namespace {
             _domain,
             path,
         }
-    }
-
-    #[allow(dead_code)]
-    pub fn get_client(&self) -> &Client {
-        self._domain.get_client()
     }
 
     pub fn to_string(&self) -> String {