+++ utils/src/lib.rs
@@ -1,0 +1,9 @@
+mod tests;
+
+pub use tests::TEST_URI_STRING;
+pub use tests::TEST_S3_URI;
+
+pub use tests::local_uri_parquet;
+pub use tests::local_uri_json;
+pub use tests::local_uri_domain;
+